Fix DisplayCardDetails.kt web view, move KatexMapping.kt and for some reason now all the mapping works lol
