+++ app/src/main/java/com/belmontCrest/cardCrafter/views/cardViews/cardDeckViews/BackCards.kt
@@ -25,7 +25,7 @@ import com.belmontCrest.cardCrafter.localDatabase.tables.NotationCard
 import com.belmontCrest.cardCrafter.localDatabase.tables.MultiChoiceCard
 import com.belmontCrest.cardCrafter.localDatabase.tables.ThreeFieldCard
 import com.belmontCrest.cardCrafter.ui.theme.GetUIStyle
-import com.belmontCrest.cardCrafter.views.miscFunctions.symbols.KaTeXWebView
+import com.belmontCrest.cardCrafter.uiFunctions.symbols.KaTeXWebView
 
 private const val line = "$$\\\\text{---}\\\\text{---}\\\\text{---}\\\\text{---}\\\\text{---}" +
         "\\\\text{---}\\\\text{---}\\\\text{---}$$"
